Added more info to scenario details
Created HTML utility class for less cluttered components. Exporting reused code from casualtySlider for scenarioDetails.
